BUG Check if overlays file exists before unlinking. (#431)

+++ cortex/freesurfer.py
@@ -279,7 +279,8 @@ def import_flat(fs_subject, patch, hemis=['lh', 'rh'], cx_subject=None,
     database.db.clear_cache(cx_subject)
     # Remove overlays.svg file (FLATMAPS HAVE CHANGED)
     overlays_file = database.db.get_paths(cx_subject)['overlays']
-    os.unlink(overlays_file)
+    if os.path.exists(overlays_file):
+        os.unlink(overlays_file)
     # Regenerate it?
